+++ app/src/main/java/com/bignerdranch/android/weather/feature_settings/data/repository/GetMyCitiesRepositoryImpl.kt
@@ -13,7 +13,7 @@ class GetMyCitiesRepositoryImpl @Inject constructor(
         myCitiesDao.getSavedCitiesAsList().map {
             MyCity(
                 cityName = it.city,
-                id = it.id
+                id = it.id.toFloat()
             )
         }
