+++ app/src/main/java/com/bignerdranch/android/weather/feature_settings/domain/model/MyCity.kt
@@ -2,5 +2,5 @@ package com.bignerdranch.android.weather.feature_settings.domain.model
 
 data class MyCity(
     val cityName: String,
-    val id: Double
+    val id: Float
 )
